refact: Move existing models to crate::models module
