+++ src/lib.rs
@@ -86,6 +86,7 @@ mod entity;
 mod errors;
 mod feature;
 pub(crate) mod metering;
+mod models;
 mod network;
 mod property;
 mod segment_evaluation;
