+++ src/metering/metering.rs
@@ -14,10 +14,9 @@
 
 use log::warn;
 
-use crate::client::feature_snapshot::FeatureSnapshot;
-use crate::client::property_snapshot::PropertySnapshot;
 use crate::metering::models::MeteringDataJson;
 use crate::metering::MeteringClient;
+use crate::models::{FeatureSnapshot, PropertySnapshot};
 use crate::network::serialization::Segment;
 use crate::utils::ThreadHandle;
 use crate::{ConfigurationId, Entity};
