+++ src/models/property_snapshot.rs
