+++ src/tests/test_get_property.rs
@@ -14,8 +14,8 @@
 
 use std::collections::HashMap;
 
-use crate::client::configuration::Configuration;
 use crate::client::{AppConfigurationClient, AppConfigurationOffline};
+use crate::models::Configuration;
 use crate::{ConfigurationProvider, Value};
 use rstest::*;
